refactor: moved test classes to same package as configuration is
